Menu complete, with optional competenceGoals and Additional filter
